Bug fix - event text appears cut in new iPhones
Root cause - When the EventViewController loads, it starts loading the webView inside viewDidLoad, and resize the webView to fit content after webViewDidFinishLoad.
In some iPhones, webViewDidFinishLoad isn't fired, causing the text to appear truncated.
Fix - When the UI loads, have the initial webView height large enough to fit the screen (causing a scroll), and during the first scroll event shrink-to-fix the webView.